pyproject.toml
[tool.poetry]
name = "daps-ui"
authors = []
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
plexapi = "^4.15.16"
tqdm = "^4.66.5"
pathvalidate = "^3.2.1"
arrapi = "^1.4.13"
pyyaml = "^6.0.2"
flask = "^3.0.3"
gunicorn = "^23.0.0"
flask-sqlalchemy = "^3.1.1"
apscheduler = "^3.10.4"
rapidfuzz = "^3.10.0"
tabulate = "^0.9.0"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
djlint = "^1.35.2"
setuptools = "^75.0.0"
pynvim = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
